# Conflicts:
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 1.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 10.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 11.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 12.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 13.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 14.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 15.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 16.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 17.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 18.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 19.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 2.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 20.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 21.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 22.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 23.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 24.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 25.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 26.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 27.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 28.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 29.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 3.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 30.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 4.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 5.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 6.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 7.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 8.prefab
#	Neo/Shadow In The Shade/Assets/05.Prefabs/New map/Dungeon 9.prefab

+++ Neo/Shadow In The Shade/Assets/01.Scripts/Player/Player.cs
@@ -154,7 +154,7 @@ public class Player : MonoBehaviour, IDamagable
 
    
 
-    public void GetHit(int damage)
+    public void GetHit(float damage)
     {
 
         if (IsDie || IsHit || playerDash.isDash || isInvincibility)
